add workaround for failed installations with pip 19.0
Today about 4 hours ago, installs of this package started to fail for
what were initially unknown reasons, with the trace:

    File "setup.py", line 20, in <module>
      from PyInstaller import __version__ as version, HOMEPATH, PLATFORM
  ImportError: No module named PyInstaller

This fix is intended to patch that issue and get the package working
again. The assumed "root cause" is the release of pip 19.0, which
happened in the same timeframe: https://pypi.org/project/pip/#history

All credit to @BurntSushi for putting together this fix!